# openapi/bar-forge-api.yaml
openapi: 3.1.1
info:
  title: BARS Engine — BAR Forge API
  description: |
    BAR → quest matching and BAR registry (Custom GPT). Set BARS_API_KEY on the server; Bearer auth.
    See docs/BAR_FORGE_API.md.

    **Game Master faces (canonical):** Shaman, Challenger, Regent, Architect, Diplomat, Sage.
    Pass optional `gameMasterFace` on match/registry so the GPT keeps sect influence; registry stores it under `metadataJson.gameMasterFace`.

    **Custom GPT:** Route each user turn through a face lens when helpful—Sage may delegate. Use `tags` and `x-gm-faces` below as behavioral hints, not separate HTTP APIs.

    **Quest catalog & clustering (v1.2):** `GET /api/quests` and `POST /api/bar-registry/cluster` accept **Bearer (BARS_API_KEY)** or, when the key is unset, a **logged-in session cookie** (same-origin). **BAR→quest links** and **campaign drafts** are **browser-session** routes (`bars_player_id` cookie), not Bearer — see tags `BarQuestLink` and `CampaignDraft`.
  version: 1.2.1
  x-gm-faces:
    shaman:
      mission: Mythic threshold — belonging, ritual space, bridge between worlds.
      gptHints: Ritualize BAR naming; prefer registry save + soft match; containment before stakes.
    challenger:
      mission: Proving ground — action, edge, lever.
      gptHints: Stress-test BAR; bias toward Show Up / stakes; smallest brave next move in metadata.
    regent:
      mission: Order — roles, rules, collective tool.
      gptHints: Classify BAR by lock/role; validate before POST registry; legible assignments.
    architect:
      mission: Blueprint — strategy, backlog, sys-admin teacher.
      gptHints: Decompose BAR; link registry ↔ quests/campaignRef; structured metadata.
    diplomat:
      mission: Weave — relational field, care, connector.
      gptHints: Reframe for I/We; enrich polarity[]; consent before persisting shared BARs.
    sage:
      mission: Integration — emergence; may wear other faces as masks.
      gptHints: Route which face leads; cross-face synthesis after match+registry; meta-routing.

servers:
  - url: https://bars-engine.vercel.app
    description: Production (no trailing slash; replace if needed)

tags:
  - name: GM_Shaman
    description: Belonging, ritual, threshold — registry & gentle matching
  - name: GM_Challenger
    description: Edge, proof, action — match with stakes; Show Up bias in prompts
  - name: GM_Regent
    description: Roles, rules — validate quest ids; ordered registry narrative
  - name: GM_Architect
    description: Structure, backlog — list registry; link BAR→quest→campaign in metadata
  - name: GM_Diplomat
    description: Relational field — polarity tags; care in copy before save
  - name: GM_Sage
    description: Routing & synthesis — default router; combine other tags in instructions
  - name: QuestCatalog
    description: Matchable quest rows (CustomBar) for matcher and GPT
  - name: BarQuestLink
    description: BAR→quest link persistence — session cookie auth (not Bearer)
  - name: CampaignDraft
    description: Braided playerArc + campaignContext drafts — session cookie auth (not Bearer)

security:
  - bearerAuth: []

paths:
  /api/match-bar-to-quests:
    post:
      operationId: matchBarToQuests
      summary: Match BAR analysis to quests (primary + up to 2 secondary)
      tags:
        - GM_Challenger
        - GM_Architect
        - GM_Sage
      x-gm-face-primary: challenger
      x-gm-face-notes: Weight prompts by face; optional gameMasterFace echoed in debug.matchedOn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchBarRequest"
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchBarResponse"
        "400":
          description: Validation error
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          description: Server error

  /api/bar-registry:
    post:
      operationId: createBarRegistryRecord
      summary: Save BAR + analysis + optional quest ids to registry
      tags:
        - GM_Shaman
        - GM_Regent
        - GM_Diplomat
      x-gm-face-primary: shaman
      x-gm-face-notes: First-class BAR persistence; merge gameMasterFace into metadataJson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarRegistryRequest"
      responses:
        "200":
          description: Created id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                required:
                  - id
        "400":
          description: Validation error or invalid quest ids
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          description: Server error
    get:
      operationId: listBarRegistryRecords
      summary: List all BAR registry rows (newest first)
      tags:
        - GM_Architect
        - GM_Sage
      x-gm-face-primary: architect
      x-gm-face-notes: Holistic catalog; metabolize BARs across the project
      parameters:
        - name: limit
          in: query
          description: Max rows when not using sample (default 100, max 500). Ignored when sample is set.
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: sample
          in: query
          description: If set, return this many rows in random order (max 500). Use for "give me random BARs to work with."
          schema:
            type: integer
            minimum: 0
            maximum: 500
        - name: includeTotal
          in: query
          description: If 1 or true, include totalCount (all rows in DB). Use for "how many BARs in the registry?"
          schema:
            type: string
            enum:
              - "0"
              - "1"
              - "true"
              - "false"
      responses:
        "200":
          description: Registry list
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/BarRegistryRecord"
                  count:
                    type: integer
                    description: Number of rows in this response
                  totalCount:
                    type: integer
                    description: Present when includeTotal=1 — total rows in bar_forge_records
                required:
                  - records
                  - count
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /api/bar-registry/{id}:
    get:
      operationId: getBarRegistryRecord
      summary: Get one registry row by id
      tags:
        - GM_Architect
        - GM_Sage
      x-gm-face-primary: sage
      parameters:
        - name: id
          in: path
          required: true
          description: Registry record CUID
          schema:
            type: string
      responses:
        "200":
          description: Single record
          content:
            application/json:
              schema:
                type: object
                properties:
                  record:
                    $ref: "#/components/schemas/BarRegistryRecord"
                required:
                  - record
        "404":
          description: Not found
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /api/quests:
    get:
      operationId: listQuestCatalog
      summary: List matchable quests (canonical CustomBar catalog slice)
      description: |
        Filterable catalog for matcher and tools. Auth: Bearer (BARS_API_KEY), or session cookie when key unset (local dev).
      tags:
        - QuestCatalog
        - GM_Architect
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: campaignRef
          in: query
          description: Filter by CustomBar.campaignRef (e.g. bruised-banana)
          schema:
            type: string
        - name: questSource
          in: query
          description: Filter by CustomBar.questSource (e.g. system, bounty)
          schema:
            type: string
      responses:
        "200":
          description: Quest rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  quests:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestDto"
                  count:
                    type: integer
                required:
                  - quests
                  - count
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/bar-registry/cluster:
    post:
      operationId: clusterBarsForArcs
      summary: Group BAR ids into arcs (manual or heuristic v1)
      description: |
        Deterministic clustering contract (spec D8). Auth: Bearer or session cookie when key unset.
      tags:
        - GM_Architect
        - GM_Sage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterRequest"
      responses:
        "200":
          description: Arc buckets + revision audit stub
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterResponse"
        "400":
          description: Validation error or unknown bar id
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/bar-quest-links:
    get:
      operationId: listBarQuestLinks
      summary: List links for a source BAR (session cookie)
      tags:
        - BarQuestLink
      parameters:
        - name: sourceBarId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Links for source BAR
          content:
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                    items:
                      $ref: "#/components/schemas/BarQuestLink"
                required:
                  - links
        "400":
          description: Missing sourceBarId
        "401":
          description: Not logged in
        "403":
          description: Not owner of source BAR
    post:
      operationId: createBarQuestLink
      summary: Create a proposed BAR→quest link (session cookie)
      tags:
        - BarQuestLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarQuestLinkCreate"
      responses:
        "200":
          description: Created link
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    $ref: "#/components/schemas/BarQuestLink"
                required:
                  - link
        "400":
          description: Validation error
        "401":
          description: Not logged in
        "403":
          description: Not owner of source BAR

  /api/bar-quest-links/{id}:
    get:
      operationId: getBarQuestLink
      summary: Get one link (session cookie)
      tags:
        - BarQuestLink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    $ref: "#/components/schemas/BarQuestLink"
                required:
                  - link
        "404":
          description: Not found
        "401":
          description: Not logged in
        "403":
          description: Forbidden
    patch:
      operationId: patchBarQuestLink
      summary: Accept, reject, or withdraw a link (steward/admin rules in description)
      description: |
        `accepted`/`rejected` require admin or instance owner/steward (InstanceMembership.roleKey owner|steward) when `instanceId` or `campaignRef` is set. `withdrawn` allowed for link author or source BAR owner.
      tags:
        - BarQuestLink
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarQuestLinkPatch"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  link:
                    $ref: "#/components/schemas/BarQuestLink"
                required:
                  - link
        "401":
          description: Not logged in
        "403":
          description: Forbidden
        "409":
          description: Invalid state transition

  /api/campaign-drafts:
    get:
      operationId: listCampaignDrafts
      summary: List current player campaign drafts (session cookie)
      tags:
        - CampaignDraft
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  drafts:
                    type: array
                    items:
                      $ref: "#/components/schemas/CampaignDraft"
                required:
                  - drafts
        "401":
          description: Not logged in
    post:
      operationId: createCampaignDraft
      summary: Create a campaign draft (session cookie)
      tags:
        - CampaignDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignDraftCreate"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft:
                    $ref: "#/components/schemas/CampaignDraft"
                required:
                  - draft
        "400":
          description: Validation error
        "401":
          description: Not logged in

  /api/campaign-drafts/{id}:
    get:
      operationId: getCampaignDraft
      summary: Get one campaign draft (session cookie)
      tags:
        - CampaignDraft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft:
                    $ref: "#/components/schemas/CampaignDraft"
                required:
                  - draft
        "404":
          description: Not found
        "401":
          description: Not logged in
        "403":
          description: Forbidden
    patch:
      operationId: patchCampaignDraft
      summary: Update a campaign draft (session cookie)
      tags:
        - CampaignDraft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignDraftPatch"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft:
                    $ref: "#/components/schemas/CampaignDraft"
                required:
                  - draft
        "404":
          description: Not found
        "401":
          description: Not logged in
        "403":
          description: Forbidden

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: Value of BARS_API_KEY from server environment

  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    ServiceUnavailable:
      description: BARS_API_KEY not configured on server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"

  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: string

    GameMasterFace:
      type: string
      description: Canonical GM sect lens (see .agent/context/game-master-sects.md)
      enum:
        - shaman
        - challenger
        - regent
        - architect
        - diplomat
        - sage

    BarAnalysis:
      type: object
      properties:
        type:
          type: string
          enum:
            - perception
            - identity
            - relational
            - systemic
        wavePhase:
          type: string
          enum:
            - Wake Up
            - Clean Up
            - Grow Up
            - Show Up
        polarity:
          type: array
          items:
            type: string
      required:
        - type
        - wavePhase
        - polarity

    MatchBarRequest:
      type: object
      properties:
        bar:
          type: string
          minLength: 1
        analysis:
          $ref: "#/components/schemas/BarAnalysis"
        gameMasterFace:
          $ref: "#/components/schemas/GameMasterFace"
        options:
          type: object
          properties:
            maxResults:
              type: integer
              minimum: 1
              maximum: 100
      required:
        - bar
        - analysis

    QuestDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        moveType:
          type: string
          nullable: true
        lockType:
          type: string
          nullable: true
        allyshipDomain:
          type: string
          nullable: true
        nation:
          type: string
          nullable: true
        archetype:
          type: string
          nullable: true
        kotterStage:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - status
        - kotterStage
        - createdAt

    MatchBarResponse:
      type: object
      properties:
        primary:
          oneOf:
            - $ref: "#/components/schemas/QuestDto"
            - type: "null"
        secondary:
          type: array
          items:
            $ref: "#/components/schemas/QuestDto"
        debug:
          type: object
          properties:
            matchedOn:
              type: array
              items:
                type: string
            confidence:
              type: number
          required:
            - matchedOn
            - confidence
      required:
        - primary
        - secondary
        - debug

    BarRegistryMatches:
      type: object
      properties:
        primaryQuestId:
          type: string
        secondaryQuestIds:
          type: array
          maxItems: 2
          items:
            type: string

    BarRegistryRequest:
      type: object
      properties:
        bar:
          type: string
          minLength: 1
        analysis:
          $ref: "#/components/schemas/BarAnalysis"
        gameMasterFace:
          $ref: "#/components/schemas/GameMasterFace"
        matches:
          $ref: "#/components/schemas/BarRegistryMatches"
        source:
          type: string
        metadataJson:
          type: object
          additionalProperties: true
      required:
        - bar
        - analysis

    BarRegistryRecord:
      type: object
      properties:
        id:
          type: string
        bar:
          type: string
        analysisType:
          type: string
        wavePhase:
          type: string
        polarity:
          type: array
          items:
            type: string
        primaryQuestId:
          type: string
          nullable: true
        secondaryQuestIds:
          type: array
          items:
            type: string
        source:
          type: string
          nullable: true
        metadataJson:
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - bar
        - analysisType
        - wavePhase
        - polarity
        - secondaryQuestIds
        - createdAt

    ClusterRequest:
      type: object
      properties:
        barIds:
          type: array
          items:
            type: string
          minItems: 1
        strategy:
          type: string
          enum:
            - manual
            - heuristic_v1
      required:
        - barIds
        - strategy

    ClusterResponse:
      type: object
      properties:
        strategy:
          type: string
        arcs:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              barIds:
                type: array
                items:
                  type: string
              reason:
                type: string
        revision:
          type: object
          properties:
            actor:
              type: string
            at:
              type: string
      required:
        - strategy
        - arcs

    BarQuestLinkCreate:
      type: object
      properties:
        sourceBarId:
          type: string
        targetQuestId:
          type: string
        matchType:
          type: string
          enum:
            - primary
            - secondary
            - manual
        confidence:
          type: number
          nullable: true
        reason:
          type: string
        supportedBy:
          type: object
          additionalProperties: true
          nullable: true
        campaignRef:
          type: string
          nullable: true
        instanceId:
          type: string
          nullable: true
      required:
        - sourceBarId
        - targetQuestId
        - reason

    BarQuestLinkPatch:
      type: object
      properties:
        status:
          type: string
          enum:
            - accepted
            - rejected
            - withdrawn
      required:
        - status

    BarQuestLink:
      type: object
      properties:
        id:
          type: string
        sourceBarId:
          type: string
        targetQuestId:
          type: string
        matchType:
          type: string
        confidence:
          type: number
          nullable: true
        reason:
          type: string
        supportedBy:
          nullable: true
        status:
          type: string
        createdByPlayerId:
          type: string
        reviewedByPlayerId:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        campaignRef:
          type: string
          nullable: true
        instanceId:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        targetQuest:
          type: object
          additionalProperties: true

    CampaignDraftCreate:
      type: object
      properties:
        title:
          type: string
          nullable: true
        playerArc:
          type: object
          additionalProperties: true
        campaignContext:
          type: object
          additionalProperties: true
        structure:
          type: object
          additionalProperties: true
      required:
        - playerArc
        - campaignContext

    CampaignDraftPatch:
      type: object
      properties:
        title:
          type: string
          nullable: true
        contentLayer:
          type: string
          enum:
            - ops
            - story
        status:
          type: string
          enum:
            - draft
            - review
            - approved
            - archived
        playerArc:
          type: object
          additionalProperties: true
        campaignContext:
          type: object
          additionalProperties: true
        structure:
          type: object
          additionalProperties: true

    CampaignDraft:
      type: object
      properties:
        id:
          type: string
        playerId:
          type: string
        title:
          type: string
          nullable: true
        contentLayer:
          type: string
          enum:
            - ops
            - story
        status:
          type: string
        playerArc:
          type: object
          additionalProperties: true
        campaignContext:
          type: object
          additionalProperties: true
        structure:
          type: object
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - playerId
        - contentLayer
        - status
        - playerArc
        - campaignContext
        - structure
        - createdAt
        - updatedAt